% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WebMasterExcluded.R
\name{WebMasterExcluded}
\alias{WebMasterExcluded}
\title{Get excluded links}
\usage{
WebMasterExcluded(token = NULL, host_id = NULL)
}
\arguments{
\item{token}{Your API token, see \code{\link{https://tech.yandex.ru/oauth/}}}

\item{host_id}{Your host, take it from URL page of the project. example "http:www.site.ru:80"}
}
\description{
Function to excluded links
}
\examples{
WebMasterExcluded()
}
