% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WebMasterExternal.R
\name{WebMasterExternal}
\alias{WebMasterExternal}
\title{Get external links}
\usage{
WebMasterExternal(token = NULL, host_id = NULL)
}
\arguments{
\item{token}{Your API token, see \code{\link{https://tech.yandex.ru/oauth/}}}

\item{host_id}{Your host, take it from URL page of the project. example "http:www.site.ru:80"}
}
\description{
Function to get external links
}
\examples{
WebMasterExternal()
}
