% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WebMasterPopular.R
\name{WebMasterPopular}
\alias{WebMasterPopular}
\title{Get popular keywords by shows and clicks for last 2 weeks}
\usage{
WebMasterPopular(token = NULL, host_id = NULL)
}
\arguments{
\item{token}{Your API token, see \code{\link{https://tech.yandex.ru/oauth/}}}

\item{host_id}{Your host, take it from URL page of the project. example "http:www.site.ru:80"}
}
\description{
Function to get popular keywords by shows and clicks for last 2 weeks
}
\examples{
WebMasterExternal()
}
